requirements.txt
python-dotenv==0.19.2
tweepy==4.12.1
s3fs==2022.11.0
